+++ src/storage/HospitalStorage.js
@@ -33,7 +33,15 @@ class HospitalStorage {
   }
 
   static async deleteHospital(id) {
-    return HospitalModel.findByIdAndDelete(id).lean();
+      const activeCasesCount = await CaseStorage.getActiveCasesCountForEntities('hospitals', id);
+    if (activeCasesCount > 0) {
+      throw new Error('Cannot delete hospital: active cases are associated with this client.');
+    }
+    const deletedHospital = await HospitalModel.findByIdAndDelete(id).lean();
+    if (!deletedHospital) {
+      throw new Error('Hospital not found or already deleted.');
+    }
+    return deletedHospital;
   }
 }
